Restore commonMain to AOSP state (move from common)
Partially fixes https://youtrack.jetbrains.com/issue/CMP-5740/Upstreaming.-compilation.-other-fixes

Restore to the last merged Jetpack Compose (1.8.0-alpha07):
- move additional code from commonMain to skikoMain. It is used only by CMP source sets. Even if it makes sense to be common, upstreaming will take a lot of time.
- implement restored expects

## Testing
- CI passes

## Release Notes
N/A
