+++ compose/animation/animation-core/src/commonMain/kotlin/androidx/compose/animation/core/Animation.kt
@@ -18,7 +18,6 @@ package androidx.compose.animation.core
 
 import androidx.annotation.RestrictTo
 import androidx.compose.animation.core.internal.JvmDefaultWithCompatibility
-import kotlin.math.roundToLong
 
 /**
  * This interface provides a convenient way to query from an [VectorizedAnimationSpec] or
@@ -88,13 +87,6 @@ internal val Animation<*, *>.durationMillis: Long
     get() = durationNanos / MillisToNanos
 
 internal const val MillisToNanos: Long = 1_000_000L
-internal const val SecondsToNanos: Long = 1_000_000_000L
-
-internal fun convertSecondsToNanos(seconds: Float): Long =
-    (seconds.toDouble() * SecondsToNanos).roundToLong()
-
-internal fun convertNanosToSeconds(nanos: Long): Double =
-    nanos.toDouble() / SecondsToNanos
 
 internal const val SecondsToMillis: Long = 1_000L
