+++ compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/lazy/layout/LazyLayoutItemAnimator.kt
@@ -126,7 +126,6 @@ internal class LazyLayoutItemAnimator<T : LazyLayoutMeasuredItem> {
                         graphicsContext,
                         layoutMinOffset,
                         layoutMaxOffset,
-                        item.crossAxisOffset,
                     )
                     keyToItemInfoMap[item.key] = newItemInfo
                     if (item.index != previousIndex && previousIndex != -1) {
@@ -154,7 +153,6 @@ internal class LazyLayoutItemAnimator<T : LazyLayoutMeasuredItem> {
                             graphicsContext,
                             layoutMinOffset,
                             layoutMaxOffset,
-                            item.crossAxisOffset,
                         )
                         itemInfo.animations.forEach { animation ->
                             if (
@@ -492,7 +490,7 @@ internal class LazyLayoutItemAnimator<T : LazyLayoutMeasuredItem> {
             graphicsContext: GraphicsContext,
             layoutMinOffset: Int,
             layoutMaxOffset: Int,
-            crossAxisOffset: Int
+            crossAxisOffset: Int = positionedItem.crossAxisOffset
         ) {
             if (!isRunningPlacement) {
                 this.layoutMinOffset = layoutMinOffset
