+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/Expect.kt
@@ -51,5 +51,3 @@ internal expect fun removePost(token: Any?)
 internal expect abstract class PlatformOptimizedCancellationException(
     message: String? = null
 ) : CancellationException
-
-internal expect fun getCurrentThreadId(): Long
