+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/Expect.desktop.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,6 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.viewinterop
+package androidx.compose.ui
 
-import androidx.compose.ui.InternalComposeUiApi
-
-/**
- * A typealias for the platform's built-in View type, which may be hosted inside of a Compose UI
- * hierarchy to allow for interoperability. Not all platforms support interoperability in this way,
- * in which case the typealias will resolve to [Any].
- */
-@InternalComposeUiApi expect class InteropView
+internal actual fun getCurrentThreadId(): Long = Thread.currentThread().id
