+++ compose/ui/ui/src/jvmMain/kotlin/androidx/compose/ui/Actual.jvm.kt
@@ -65,4 +65,3 @@ internal actual abstract class PlatformOptimizedCancellationException actual con
 
 }
 
-internal actual fun getCurrentThreadId(): Long = Thread.currentThread().id
