+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/viewinterop/InteropView.skiko.kt
@@ -195,3 +195,9 @@ private fun <T : InteropView> LayoutNode.requireViewFactoryHolder(): TypedIntero
     return checkNotNull(interopViewFactoryHolder) as TypedInteropViewHolder<T>
 }
 
+/**
+ * A typealias for the platform's built-in view group type. It's used as a wrapper of [InteropView]
+ * and might have different from [InteropView] type. Not all platforms support interoperability
+ * in this way, in which case the typealias will resolve to [Any].
+ */
+internal expect class InteropViewGroup
